Add flux on temp management cluster and refactor flux install on all clusters to use CRS, replace Calico with Cilium (#8)
refactor: replace Calico with Cilium and install Flux on all clusters using ClusterResourceSet

+++ scripts/brutal-aws-cleanup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+echo EXIT.
+echo Read the script and uncomment exit statement below.
 exit 1
 
 # Quick delete all resources that make up the clusters provisioned in this account. (did I say it is brutal)
@@ -8,9 +10,11 @@ exit 1
 set -eoux pipefail
 
 running_instances=$(aws ec2 describe-instances --filters "Name=instance-state-name,Values=running" --query "Reservations[].Instances[].InstanceId" --output=text)
-aws ec2 terminate-instances --instance-ids $running_instances
+if [ -n "$running_instances" ]; then
+  aws ec2 terminate-instances --instance-ids $running_instances
+fi
 
-nat_gateways=$(aws ec2 describe-nat-gateways --query "NatGateways[].NatGatewayId")
+nat_gateways=$(aws ec2 describe-nat-gateways --query "NatGateways[].NatGatewayId" --output=text)
 for n in ${nat_gateways[@]}; do
   aws ec2 delete-nat-gateway --nat-gateway-id $n
 done
@@ -18,7 +22,14 @@ done
 aws elb delete-load-balancer --load-balancer-name=mgmt-apiserver
 sleep 60
 
-aws ec2 release-address --allocation-id=$(aws ec2 describe-addresses --query "Addresses[].AllocationId" --output=text)
+eips=$(aws ec2 describe-addresses --query "Addresses[].AllocationId" --output=text)
+for i in ${eips[@]}; do
+  aws ec2 release-address --allocation-id $i
+done
+
+echo Done.
+
+echo You may need to delete VPC manually from the console
 
 # Some time after NAT and LB are deleted and ENIs are deleted it will be possible to simply delete VPC from the console.
 # The command will still hang though, if attempted to delete from CLI:
